requirements.txt
langchain
langchain-community
langchain-openai
chromadb
sentence-transformers
datasets
python-dotenv
openai
pandas
matplotlib
seaborn
